requirements.txt
# Core Dependencies (Production-Grade Backend)
# Minimal set of dependencies following clean architecture principles

# Minimal production requirements for ai-engine

fastapi==0.115.0
uvicorn
pydantic==2.9.0
python-multipart==0.0.12

openai==1.54.0
langchain==0.3.15
langchain-openai==0.2.11
langchain-core==0.3.31
langchain-community==0.3.15
langgraph==0.2.28
redis[hiredis]>=5.0.1

faiss-cpu==1.13.2  # updated: 1.9.0.post1 unavailable on PyPI for this platform
langchain-text-splitters==0.3.3
graphviz==0.20.3

pymongo==4.8.0
python-jose[cryptography]==3.3.0

pypdf==5.0.0
python-docx==1.1.2

python-dotenv==1.0.0
requests==2.32.4
httpx==0.27.2

langdetect==1.0.9

nest-asyncio==1.6.0


python-jose[cryptography]
PyPDF2==3.0.1